Rucksack Reorganization — Part 2
This one actually shocked me the way it came out terse. This is a true
testament to leveraging the language constructs to write cleaner, "more
literate" code. — "literate" in the sense of "Literate Programming" now
(google it), NOT "literate" as an opposite of "illiterate".

Pseudocode…
- Combine the characters of each line in the group with "intersection"
- Add the priority of the similar character to our list of priorities
- The answer is the sum of priorities

Couldn't be happier with the way this came out.