Camp Cleanup — Part 2
For this one, I know it's highly inefficient, and should not be used for
large arrays. I settled for this because it seemed the ranges were
between [0, 100], and the number of test cases was 1000. Anything more
than 10000 test cases with ranges possibly on the order of thousands may
begin to crash this memory-wise.

A more efficient way would be to check the bounds. It's faster, it
doesn't require more memory space, it's leaner. However, I chose to
trade that off for literacy. The idea is that someone who isn't too
technical should be able to read this and understand what's going on.

🤞🏾